Fixed issue found while testing Windows
