+++ .github/README.md
@@ -108,7 +108,7 @@ curl -sf -L https://raw.githubusercontent.com/meshcloud/collie-cli/main/install.
 
 **Windows**
 
-Just copy the content of [`install.ps1`](https://github.com/meshcloud/collie-cli/blob/develop/install.ps1) and run it in your PowerShell console.
+Simply copy the content of [`install.ps1`](https://github.com/meshcloud/collie-cli/blob/develop/install.ps1) and run it in your PowerShell console.
 
 ## 👋 Need help or have feedback?
