+++ install.sh
@@ -20,7 +20,7 @@ case "$(uname -s)" in
      name="collie-x86_64-unknown-linux-gnu"
      ;;
    CYGWIN*|MINGW32*|MSYS*|MINGW*)
-     echo 'Please use install.ps1 script. See readme for more information.'
+     echo 'Please execute the install.ps1 script as mentioned in our README. Go to https://github.com/meshcloud/collie-cli#-install-and-usage'
      exit 1
      ;;
    *)
