+++ src/presentation/mesh-table.ts
@@ -1,6 +1,7 @@
 import { MeshTag } from "../mesh/mesh-tenant.model.ts";
 import { QueryStatistics } from "../mesh/query-statistics.ts";
 import { dim, Table, yellow } from "../deps.ts";
+import { CLICommand } from "../config/config.model.ts";
 
 export abstract class TableGenerator {
   abstract getColumns(): string[];
@@ -21,8 +22,29 @@ export abstract class TableGenerator {
   }
 }
 
-export interface MeshTable {
-  draw(generator: TableGenerator, stats: QueryStatistics | null): void;
+export abstract class MeshTable {
+  abstract draw(generator: TableGenerator, stats: QueryStatistics | null): void;
+
+  protected formatStats(stats: QueryStatistics): string {
+    const summary: string[] = [];
+
+    if (stats.duration["cache"]) {
+      summary.push(
+        `Loaded from cache in ${stats.duration.cache}ms. See "${CLICommand} cache" for details.`,
+      );
+    }
+
+    const cloudstats = Object
+      .entries(stats.duration)
+      .filter(([cloud]) => cloud !== "cache")
+      .map(([cloud, d]) => `${cloud}: ${(d || 0) / 1000}s`);
+
+    if (cloudstats.length) {
+      summary.push(`Queried from cloud - ${cloudstats.join(", ")}.`);
+    }
+
+    return summary.join(" ");
+  }
 }
 
 export class MeshTableTag implements MeshTag {
