+++ src/presentation/no-tty-mesh-table.ts
@@ -1,9 +1,10 @@
 import { log, Table } from "../deps.ts";
 import { bold } from "../deps.ts";
 import { MeshTable, TableGenerator } from "./mesh-table.ts";
+import { QueryStatistics } from "../mesh/query-statistics.ts";
 
-export class NoTtyMeshTable implements MeshTable {
-  draw(generator: TableGenerator): void {
+export class NoTtyMeshTable extends MeshTable {
+  draw(generator: TableGenerator, stats: QueryStatistics | null): void {
     const rows = generator.getRows();
 
     if (rows.length === 0) {
@@ -16,5 +17,11 @@ export class NoTtyMeshTable implements MeshTable {
       .header(generator.getColumns().map((value) => bold(value)))
       .body(rows)
       .render();
+
+    if (stats) {
+      console.log(
+        this.formatStats(stats),
+      );
+    }
   }
 }
